requirements.txt
certifi==2024.8.30
charset-normalizer==3.3.2
colorama==0.4.6
commonmark==0.9.1
idna==3.10
markdown-it-py==3.0.0
mdurl==0.1.2
packaging==24.1
pip-review==1.3.0
prompt_toolkit==3.0.36
Pygments==2.18.0
PyYAML==6.0.2
qrcode==8.0
questionary==1.10.0
requests==2.32.3
rich==13.9.2
urllib3==2.2.3
wcwidth==0.2.13
